+++ src/Servers/Connections.Abstractions/src/IMulitplexedConnectionListener.cs
@@ -7,12 +7,12 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http.Features;
 
-namespace Microsoft.AspNetCore.Connections
+namespace Microsoft.AspNetCore.Connections.Experimental
 {
     /// <summary>
     /// Defines an interface that represents a listener bound to a specific <see cref="EndPoint"/>.
     /// </summary>
-    public interface IMultiplexedConnectionListener : IAsyncDisposable
+    internal interface IMultiplexedConnectionListener : IAsyncDisposable
     {
         /// <summary>
         /// The endpoint that was bound. This may differ from the requested endpoint, such as when the caller requested that any free port be selected.
