+++ src/Servers/Connections.Abstractions/src/IMultiplexedConnectionListenerFactory.cs
@@ -6,12 +6,12 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http.Features;
 
-namespace Microsoft.AspNetCore.Connections
+namespace Microsoft.AspNetCore.Connections.Experimental
 {
     /// <summary>
     /// Defines an interface that provides the mechanisms for binding to various types of <see cref="EndPoint"/>s.
     /// </summary>
-    public interface IMultiplexedConnectionListenerFactory
+    internal interface IMultiplexedConnectionListenerFactory
     {
         /// <summary>
         /// Creates an <see cref="IMultiplexedConnectionListener"/> bound to the specified <see cref="EndPoint"/>.
