+++ src/Servers/Connections.Abstractions/src/MultiplexedConnectionBuilder.cs
@@ -6,9 +6,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Microsoft.AspNetCore.Connections
+namespace Microsoft.AspNetCore.Connections.Experimental
 {
-    public class MultiplexedConnectionBuilder : IMultiplexedConnectionBuilder
+    internal class MultiplexedConnectionBuilder : IMultiplexedConnectionBuilder
     {
         private readonly IList<Func<MultiplexedConnectionDelegate, MultiplexedConnectionDelegate>> _components = new List<Func<MultiplexedConnectionDelegate, MultiplexedConnectionDelegate>>();
