+++ src/Servers/Connections.Abstractions/src/MultiplexedConnectionDelegate.cs
@@ -3,12 +3,12 @@
 
 using System.Threading.Tasks;
 
-namespace Microsoft.AspNetCore.Connections
+namespace Microsoft.AspNetCore.Connections.Experimental
 {
     /// <summary>
     /// A function that can process a connection.
     /// </summary>
     /// <param name="connection">A <see cref="MultiplexedConnectionContext" /> representing the connection.</param>
     /// <returns>A <see cref="Task"/> that represents the connection lifetime. When the task completes, the connection will be closed.</returns>
-    public delegate Task MultiplexedConnectionDelegate(MultiplexedConnectionContext connection);
+    internal delegate Task MultiplexedConnectionDelegate(MultiplexedConnectionContext connection);
 }
