+++ src/Servers/Kestrel/Core/src/Internal/Http3ConnectionContext.cs
@@ -4,6 +4,7 @@
 using System.Buffers;
 using System.Net;
 using Microsoft.AspNetCore.Connections;
+using Microsoft.AspNetCore.Connections.Experimental;
 using Microsoft.AspNetCore.Http.Features;
 using Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Infrastructure;
