+++ src/Servers/Kestrel/Transport.Quic/src/QuicConnectionFactory.cs
@@ -9,6 +9,7 @@ using System.Net.Security;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Connections;
+using Microsoft.AspNetCore.Connections.Experimental;
 using Microsoft.AspNetCore.Http.Features;
 using Microsoft.AspNetCore.Server.Kestrel.Transport.Experimental.Quic.Internal;
 using Microsoft.Extensions.Logging;
@@ -16,7 +17,7 @@ using Microsoft.Extensions.Options;
 
 namespace Microsoft.AspNetCore.Server.Kestrel.Transport.Experimental.Quic
 {
-    public class QuicConnectionFactory : IMultiplexedConnectionFactory
+    internal class QuicConnectionFactory : IMultiplexedConnectionFactory
     {
         private QuicTransportContext _transportContext;
