+++ src/Servers/Kestrel/Transport.Quic/src/WebHostBuilderMsQuicExtensions.cs
@@ -3,6 +3,7 @@
 
 using System;
 using Microsoft.AspNetCore.Connections;
+using Microsoft.AspNetCore.Connections.Experimental;
 using Microsoft.AspNetCore.Server.Kestrel.Transport.Experimental.Quic;
 using Microsoft.Extensions.DependencyInjection;
