+++ src/Servers/Kestrel/Transport.Sockets/src/Client/SocketConnectionFactory.cs
@@ -16,7 +16,7 @@ using Microsoft.Extensions.Options;
 
 namespace Microsoft.AspNetCore.Server.Kestrel.Transport.Sockets
 {
-    public class SocketConnectionFactory : IConnectionFactory, IAsyncDisposable
+    internal class SocketConnectionFactory : IConnectionFactory, IAsyncDisposable
     {
         private readonly SocketTransportOptions _options;
         private readonly MemoryPool<byte> _memoryPool;
