--- src/Servers/Kestrel/samples/QuicSampleApp/QuicSampleApp.csproj
@@ -1,27 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>$(DefaultNetCoreTargetFramework)</TargetFramework>
-    <NoDefaultLaunchSettingsFile>true</NoDefaultLaunchSettingsFile>
-    <GenerateRazorAssemblyInfo>false</GenerateRazorAssemblyInfo>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Reference Include="Microsoft.AspNetCore.Server.Kestrel" />
-    <Reference Include="Microsoft.Extensions.Logging.Console" />
-    <Reference Include="Microsoft.AspNetCore.Server.Kestrel.Transport.Experimental.Quic" />
-  </ItemGroup>
-
-  <ItemGroup Condition="'$(OS)' == 'Windows_NT'">
-    <PackageReference Include="MsQuicPackage.win-x64" Version="1.0.9">
-      <AllowExplicitReference>true</AllowExplicitReference>
-    </PackageReference>
-  </ItemGroup>
-
-  <ItemGroup Condition="'$(OS)' == 'Unix'">
-    <PackageReference Include="MsQuicPackage.linux-x64" Version="1.0.9">
-      <AllowExplicitReference>true</AllowExplicitReference>
-    </PackageReference>
-  </ItemGroup>
-
-</Project>
