+++ src/Servers/Kestrel/shared/TransportMultiplexedConnection.cs
@@ -6,6 +6,7 @@ using System.Collections.Generic;
 using System.IO.Pipelines;
 using System.Net;
 using System.Threading;
+using Microsoft.AspNetCore.Connections.Experimental;
 using Microsoft.AspNetCore.Http.Features;
 
 namespace Microsoft.AspNetCore.Connections
